move other utilities internal
